src/services/agents_catalogue/genspec/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
pyyaml>=6.0
jsonschema>=4.17.3
markdown>=3.4.3
python-dotenv>=1.0.0
requests>=2.31.0
python-markdown-math>=0.8
pygments>=2.15.1
inquirer>=3.1.3
langchain>=0.0.267
langchain-community>=0.0.10 